Moved MessageCopy to it's own data objects file
