+++ ai/data_objects.py
@@ -1,0 +1,10 @@
+class MessageCopy:
+    def __init__(
+        self,
+        content=None,
+        display_name=None,
+        avatar_url=None
+    ):
+        self.content = content
+        self.display_name = display_name
+        self.avatar_url = avatar_url
