+++ db/data_objects.py
@@ -15,18 +15,6 @@ def map_to_object(row, constructor):
     return constructor(**row)
 
 
-class MessageCopy:
-    def __init__(
-        self,
-        content=None,
-        display_name=None,
-        avatar_url=None
-    ):
-        self.content = content
-        self.display_name = display_name
-        self.avatar_url = avatar_url
-
-
 class Drone:
 
     def __init__(
